+++ examples/with-preferred-language-middleware/src/features/common/components/Layout.tsx
@@ -2,9 +2,14 @@ import type { RouteLocale } from 'next-roots'
 import type { PropsWithChildren, ReactNode } from 'react'
 import { Nav } from './Nav'
 
-import Link from 'next/link'
 import 'src/features/common/styles.css'
-import { getAboutHref, getContactsHref, getHomeHref } from 'src/server/router'
+import {
+  getAboutHref,
+  getBooksHref,
+  getContactsHref,
+  getHomeHref,
+  getProductsHref,
+} from 'src/server/router'
 import { getDictionary } from 'src/server/utils/getDictionary'
 import { Footer } from './Footer'
 
@@ -17,6 +22,8 @@ async function getNavigation(locale: string) {
   const t = await getDictionary(locale)
   return [
     { name: t('nav.Home'), href: getHomeHref(locale) },
+    { name: t('nav.Products'), href: getProductsHref({ locale }) },
+    { name: t('nav.Books'), href: getBooksHref(locale) },
     { name: t('nav.About'), href: getAboutHref(locale) },
     { name: t('nav.Contacts'), href: getContactsHref(locale) },
   ]