+++ examples/with-preferred-language-middleware/src/routes/products/i18n.js
@@ -1,0 +1,5 @@
+module.exports.routeNames = [
+  { locale: 'en', path: 'products' },
+  { locale: 'cs', path: 'produkty' },
+  { locale: 'es', path: 'productos' },
+]